backend/requirements.txt
fastapi
python-multipart
uvicorn
python-dotenv
httpx
websockets
requests
pydantic
pypinyin
openai
